feat: arrays util
