+++ src/commands/settings/enablecategory.ts
@@ -1,7 +1,7 @@
 import { Guild, Message } from 'discord.js';
 import { GenericCommand } from '#lib/structures/commands';
 import { Identifiers, UserError } from '#lib/errors';
-import { removeDuplicates } from '#utils/util';
+import { removeDuplicates } from '#utils/functions';
 
 abstract class EnableCategoryCommand extends GenericCommand {
 	constructor() {
