+++ src/lib/utils/functions/index.ts
@@ -1,3 +1,4 @@
 export * from './numbers.js';
 export * from './permissions.js';
 export * from './strings.js';
+export * from './arrays.js';
